+++ src/main/java/um/tesoreria/core/kotlin/model/Bancaria.kt
@@ -8,6 +8,7 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
+import um.tesoreria.core.hexagonal.cuenta.infrastructure.persistence.entity.CuentaEntity
 import java.math.BigDecimal
 
 @Entity
@@ -42,6 +43,6 @@ data class Bancaria(
 
     @OneToOne(optional = true)
     @JoinColumn(name = "cub_pla_cuenta", insertable = false, updatable = false)
-    var cuenta: Cuenta? = null,
+    var cuenta: CuentaEntity? = null,
 
     ) : Auditable()