+++ src/main/java/um/tesoreria/core/kotlin/model/Dependencia.kt
@@ -1,6 +1,7 @@
 package um.tesoreria.core.kotlin.model
 
 import jakarta.persistence.*
+import um.tesoreria.core.hexagonal.cuenta.infrastructure.persistence.entity.CuentaEntity
 import um.tesoreria.core.hexagonal.geografica.infrastructure.persistence.entity.GeograficaEntity
 import java.math.BigDecimal
 
@@ -26,7 +27,7 @@ data class Dependencia(
 
     @OneToOne(optional = true)
     @JoinColumn(name = "cuentaHonorariosPagar", insertable = false, updatable = false)
-    val cuenta: Cuenta? = null
+    val cuenta: CuentaEntity? = null
 
 ) : Auditable() {
 
@@ -55,7 +56,7 @@ data class Dependencia(
         var cuentaHonorariosPagar: BigDecimal? = null
         var facultad: Facultad? = null
         var geografica: GeograficaEntity? = null
-        var cuenta: Cuenta? = null
+        var cuenta: CuentaEntity? = null
 
         fun dependenciaId(dependenciaId: Int?) = apply { this.dependenciaId = dependenciaId }
         fun nombre(nombre: String) = apply { this.nombre = nombre }
@@ -65,7 +66,7 @@ data class Dependencia(
         fun cuentaHonorariosPagar(cuentaHonorariosPagar: BigDecimal?) = apply { this.cuentaHonorariosPagar = cuentaHonorariosPagar }
         fun facultad(facultad: Facultad?) = apply { this.facultad = facultad }
         fun geografica(geografica: GeograficaEntity?) = apply { this.geografica = geografica }
-        fun cuenta(cuenta: Cuenta?) = apply { this.cuenta = cuenta }
+        fun cuenta(cuenta: CuentaEntity?) = apply { this.cuenta = cuenta }
 
         fun build() = Dependencia(this)
     }